+++ package.json
@@ -10,8 +10,8 @@
     "distro": "run-s build test:build",
     "lint": "eslint .",
     "prepublishOnly": "run-s distro",
-    "dev": "mocha -r esm --reporter=spec --recursive --watch test/spec",
-    "test": "nyc --reporter=lcov mocha -r esm --reporter=spec --recursive test/spec",
+    "dev": "mocha -r esm --reporter=spec --recursive --watch test/**/*Spec.js",
+    "test": "nyc --reporter=lcov mocha -r esm --reporter=spec --recursive test/**/*Spec.js",
     "test:build": "mocha --reporter=spec --recursive test/distro",
     "all": "run-s lint test distro"
   },
