+++ src/platform/extractors/extractEventDefinitionVariables.js
