+++ src/platform/extractors/extractFormFields.js
