+++ src/platform/extractors/extractOutMappings.js
