+++ src/platform/extractors/extractOutputParameters.js
