+++ src/platform/extractors/extractResultVariables.js
