+++ src/platform/util/ElementsUtil.js
