+++ src/platform/util/EventDefinitionsUtil.js
