+++ src/platform/util/ExtensionElementsUtil.js
