+++ src/platform/util/ProcessVariablesUtil.js
