+++ test/platform/fixtures/call-activity-local.bpmn
