+++ test/platform/fixtures/call-activity.bpmn
