+++ test/platform/fixtures/error-code-variable.bpmn
