+++ test/platform/fixtures/error-message-variable.bpmn
