+++ test/platform/fixtures/result-variable.bpmn
