+++ test/platform/fixtures/sub-process-output-mapping.bpmn
