+++ test/platform/fixtures/sub-process-own-scope.bpmn
